extension: Add initial openvscode-server support
Document everything that is currently known about this introduction, in
README as usual for this extension. Main goal being, to be able to
remotely deploy this extension, alongside still locally too.

Downgrade the vscode version from 1.78.0 to 1.77.3, the latter being the
currently latest stable (Verified, [1]) one for this server integration.

This is because [1]'s version and this 'vscode' version have to align,
in order for openvscode-server to work.

Contributes to fixing the foreign (companion's) Issue [2], below.

[1]https://github.com/gitpod-io/openvscode-server/tags
[2]https://github.com/eclipse-cdt-cloud/vscode-trace-extension/issues/15

Signed-off-by: Marco Miller <marco.miller@ericsson.com>