[TENTATIVE - DO NOT MERGE] Adjust to renaming of vscode trace extension
Note: this is tentative for now. See:
https://github.com/eclipse-cdt-cloud/vscode-trace-extension/pull/213

The trace viewer extension (vscode-trace-extension) might need to be renamed,
tentatively to "vscode-trace-viewer", in orfer for it to be publishable to
the Visual Studio Marketplace. If that happens, references to the extension,
will neeed to follow suite.

This extension here uses the API offered by the trace viewer extension, and
in that eventuality will need to refer to it using the new name.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>