add ovsx as a devDependency
oxsv is needed to publish, to open vsx, the vscode extension built
from sources in this repo.

We also make sure dependency @vscode/vsce we pull is a not too recent
version, since those add a binary file that's under a proprietary
license, that's not yet approved for use in our project.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>