[license check] change batch size from 50 to 500
The batch size is used by dash-licenses, and determines how many
dependencies are sent per request to the Eclipse Foundation and
ClearlyDefined servers. 500 is actually the default when running
plain dash-licenses, and will result in fewer requests which could
help with HTTP 429 errors we see once in a while, specially on the
ClearlyDefined requests.

See:
https://github.com/eclipse-dash/dash-licenses/issues/301#issuecomment-2704821657

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>